refactor(FrontController): 최전방에서 게임 시작에 대한 입력, 예외 핸들링 역할만을 수행하는 컨트롤러로 역할 분리
- Controller 를 FrontController로 이름 변경
- try ~ catch 문에 있는 로직을 SubwayController 에게 위임
